fix: add all functions and types
This expands out to add (nearly) all the functions and types for Raylib
5.0.0 (the version I have installed via homebrew, and that's documented
on raylib.com as of this commit)

A lot of things were generated quickly so there may be some
discrepencies between the code and the underlying library. I'm still not
sure how to write tests for this.