Using an enum custom type to declare string constants
This is not yet working - the enum constructor expects the occasional
aligning value, then a list of strings which it increments the value
for. I tried coercing all the values to arrayrefs with pair, but the
constructor is unhappy with this.

It should at least show off how this *should* work. The parameter to
your FFI bound function should now be a string, e.g. 'KEY_A', and
functions can also be coerced into returning string constants - see
return type for `GetKeyPressed`.

It's not a radical departure API-wise - rlt-01.pl should reflect the
required changes